refactor: remove unused import and update navigation screens names in Navigation.js - removed import FriendScreen - renamed "findFriends" to "myFriend" for component name in navigation stack + added new imports for SuggestedFriends, SentRequest, ReceivedRequest, MyFriends screens + updated corresponding screen names in the navigation stack accordingly.
In Navigation.js file, I refactored the code by removing an unused import and updating the navigation screen names. I removed the import statement for FriendScreen and renamed the "findFriends" name to "myFriend" for the component in the navigation stack. Additionally, I added new imports for SuggestedFriends, SentRequest, ReceivedRequest, and MyFriends screens and updated their corresponding names in the navigation stack accordingly.