Added support for sending multiple friends' IDs and tab index to upload media and documents in helper.js
Changes:
- Removed hardcoded MediaPath from uploadChatMedia, uploadDoc functions
- Added new parameters 'tabIndex' and 'groupMemberId' to uploadChatMedia, uploadDoc functions
- Updated sendMessage function to include 'IsFriendCircle' and 'FriendsId' in formData based on tabIndex value
- Modified handleImageResponse and handleDocResponse functions to pass 'tabIndex' and 'groupMemberId' as arguments.

Reason:
To enable sending media and documents to multiple friends in a single chat tab, the helper.js file was updated to accept and process the new parameters 'tabIndex' and 'groupMemberId'. This will allow the application to send media and documents to multiple friends in the same tab without requiring separate API calls for each friend.