It looks like the given code snippet is an excerpt from an Android Studio project file, specifically the `AndroidManifest.xml` and `build.gradle`. The changes made to the files include adding more permissions in the manifest file and updating the Gradle build script with Google Services plugin and Firebase Messaging library dependencies.
The new permissions added in the manifest file include:
- INTERNET
- ACCESS_NETWORK_STATE
- READ_EXTERNAL_STORAGE
- WRITE_EXTERNAL_STORAGE
- RECEIVE
- WAKE_LOCK
- READ_PHONE_STATE
- SYSTEM_ALERT_WINDOW
- ACTION_MANAGE_OVERLAY_PERMISSION
- FOREGROUND_SERVICE
- CALL_PHONE
- CAMERA
- MODIFY_AUDIO_SETTINGS
- RECORD_AUDIO
- BIND_TELECOM_CONNECTION_SERVICE
- CAPTURE_VIDEO_OUTPUT
- FOREGROUND_SERVICE_MEDIA_PROJECTION

The Gradle build script was updated to include the Google Services plugin and Firebase Messaging library dependencies. The compileSdkVersion, targetSdkVersion were also updated to 33.

Additionally, a new activity named `MainActivity` with launcher intent filter was added in the manifest file. This activity might be used as the main entry point of the application.

The strings.xml resource file was also updated to include two new strings for screen capture notifications.