refactor(android/app/google-services.json): update project number, project id, storage bucket, and package name for Google Services configuration file to reflect new app details perf(android/app/google-services.json): improve security by rotating the API key in the Google Services configuration file
 It looks like this is a configuration file written in the Xcode Project format (`.xcworkspace` or `.xcodeproj`). The file contains information about the project's structure, build settings, and file references.

In particular, this excerpt shows the definition of several variables such as `files`, `baseConfigurationReference`, and `buildSettings`. These variables contain references to specific files in the project (e.g., `ringtone.mp3`, `GoogleService-Info 2.plist`, etc.) and configuration settings for the build process.

Additionally, there is a new entry added in the `buildSettings` under the key `DEVELOPMENT_TEAM` with the value `6L85YUY8GF` for the release scheme and `8LR99G7A89` for the test scheme. This could be related to code signing or provisioning profiles for the app's distribution.

Overall, this file is essential for Xcode to build and manage your iOS project correctly.

 add(ios/GoogleService-Info.plist): initialize Firebase configuration for iOS app by adding GoogleService-Info.plist file with project details and credentials.