It looks like you are working with a React Native codebase and have made some adjustments to the code by adding or removing whitespaces, changing variable names from `prevState` to `state`, and updating import statements for ES6 syntax. Here's a summary of the changes:
1. In the `componentDidUpdate` method, you changed `this.setState({ isMeetingActive: false });` to `this.setState({isMeetingActive: false});`.
2. You updated some variable names from `prevState` to just `state` in the arrow functions for `handleAttendeeJoin`, `handleAttendeeLeave`, and other methods.
3. In the import statement for `handleAttendeeJoin`, you changed `attendeeId in attendeeNameMap` to `{attendeeId} in attendeeNameMap`.
4. You updated some imports to use ES6 syntax, such as `import React from 'react';` and `import { Alert } from 'react-native';`.
5. In the `handleAttendeeLeave` method, you changed `prevState.attendees.filter` to `this.setState(prevState => ({ attendees: prevState.attendees.filter })`.
6. You added a semicolon at the end of the line `console.log('Back button pressed');`.
7. In the `handleAttendeeJoin` method, you changed `attendeeNameMap[attendeeId] = externalUserId.split('#')[1];` to `attendeeNameMap[attendeeId] = externalUserId.split('#')[1];`.
8. You added a space before the opening parenthesis in some method definitions, such as `handleAttendeeJoin = ({ attendeeId, externalUserId }) => {`.
9. In the `startOneMinuteTimer` method, you changed `this.setState({ meetingDuration: Date.now() - this.state.startTime });` to `this.setState({meetingDuration: Date.now() - this.state.startTime});`.
10. You added a space before the opening curly brace in some methods, such as `handleAttendeeJoin = ({ attendeeId, externalUserId }) => {`.

These changes should not affect the functionality of your codebase significantly, but they make it more readable and consistent with modern coding conventions.

 It looks like this code snippet is written in React Native and is a part of a component called `GroupVideoMeeting`. The component renders a video meeting interface with features such as mute, camera toggle, speaker switch, and attendee list.

The changes made to the code include:

1. Changing the import statement for the `BackHandler` module to use template literals.
2. Updating the event listener registration to use template literals.
3. Removing the release of the sound file in some places due to the absence of the corresponding initialization.
4. Updating the import statement for the `GroupAttendeeItem` component instead of the `AttendeeItem` component when rendering the attendee list.
5. Adding a new prop `isSpeakerActive` to the `SwitchMicrophoneToSpeakerButton` component.

The code still initializes the sound file and plays it when there are two or more attendees in the meeting, but it no longer releases the sound file when there is only one attendee or when the component unmounts. This could potentially lead to memory leaks if the sound file is not properly released. It's recommended to release the sound file when it's no longer needed to avoid such issues.

 It looks like you are working on a React Native component for a video meeting application. In the given code snippet, you have made some changes to the UI of the component, specifically adding and removing some components related to screen sharing.

The code starts by importing necessary modules and defining the `styles` object using `StyleSheet.create()`. The main component is defined as `VideoMeeting`, which extends `React.Component`. Inside the render method of this component, you are returning a JSX structure that includes various views and text components.

The UI structure includes a `SafeAreaView` with a `View` containing two videos (local and remote), an attendee list, and a button container with several buttons for mute, microphone to speaker, camera, switching cameras, screen sharing, and hanging off the call.

You have made some changes to this structure by adding and removing components related to screen sharing. Specifically, you have added a `ShareScreenBtn` component in the button container and removed the `smallWrapper` and `smallVideo` components.

Overall, it seems that the code is intended to display two videos (local and remote) in a video meeting application with some additional features like screen sharing and an attendee list. The changes you have made update the UI to reflect the addition of screen sharing functionality.