This code is written in React Native and uses various components from the "react-native" and "react-native-elements" libraries. The main functionality of this code is to display a list of users, which can be either friends or search results. The user list is displayed inside a modal that can be opened by pressing a button.
The state variables `searchButtonClicked`, `friendList`, and `searchResults` are used to determine whether the list should display friends or search results. The `isLoading` variable is used to show a loading shimmer effect while data is being fetched.

When a user is selected from the list, the `toggleSendBtn` function is called with the index of the selected user and a boolean value indicating whether search results are currently being displayed. This function sets the `sendBtn` property of the selected user in the state to indicate whether the "Send" or "Undo" button should be shown for that user.

The `FlatList` component is used to efficiently render a list of items, and the `TouchableOpacity` component is used to create pressable areas for each user in the list. The `Image`, `Text`, and `Modal` components are also used to display the user's profile picture, name, and the modal window, respectively.

Overall, this code creates a user interface for selecting friends or search results and allows users to send friend requests or undo previously sent requests.